+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Display additional locker location information on the map in checkout.
 - Select endorsement (abandon, return) for international, undeliverable shipments.
 
+### Fixed
+
+- Improve handling of apartment addresses when splitting the recipient street.
+
 ## 2.6.2
 
 ### Fixed